// packages/spot-ui/src/order-form/types.ts
export interface CalculatedAmount {
  raw: string;
  ui: string;
  usd: string;
}

export interface CalculatedOrderInputAmount extends CalculatedAmount {
  isEmpty: boolean;
}

export interface CalculatedMarketPriceValues {
  raw: string;
  ui: string;
  usd: string;
}

export interface CalculatedOrderSchedule {
  totalTrades: number;
  fillDelay: TimeDuration;
  fillDelayMillis: number;
  duration: TimeDuration;
  durationMillis: number;
}

export interface CalculatedPriceInput {
  raw: string;
  typedValue?: string;
  percentage: string;
  isTypedValue: boolean;
}

export interface CalculatedPriceDisplay {
  raw: string;
  ui: string;
  usd: string;
}

export interface CalculatedTriggerPriceValues extends CalculatedPriceInput {
  enabled: boolean;
  display: CalculatedPriceDisplay;
  error?: InputError;
}

export interface CalculatedLimitPriceValues extends CalculatedPriceInput {
  display: CalculatedPriceDisplay;
  error?: InputError;
}

export interface CalculatedOrderValues extends CalculatedOrderSchedule {
  orderType: OrderType;
  isMarketOrder: boolean;
  isTriggerPrice: boolean;
  slippageBps: number;
  inputAmount: string;
  outputAmount: string;
  inputAmountPerTrade: string;
  minOutputAmountPerTrade: string;
  minOutputAmountTotal: string;
  triggerOutputAmountPerTrade: string;
  tradePrice: string;
  marketPrice: string;
  limitPrice: string;
  triggerPrice: string;
  displayFeeAmount: string;
  displayFeePercent: number;
}

export interface CalculatedOrderTrades {
  totalTrades: number;
  maxTrades: number;
  inputAmountPerTrade: CalculatedAmount;
  minOutputAmountPerTrade: CalculatedAmount;
  triggerOutputAmountPerTrade: CalculatedAmount;
  error?: InputError;
}

export interface CalculatedOrderFees extends CalculatedAmount {
  percentage: number;
}

export interface CalculatedOrderFormSchedule extends CalculatedOrderSchedule {
  fillDelayError?: InputError;
  durationError?: InputError;
}

export interface CalculatedOrderFormErrors {
  primary?: InputError;
  all: InputError[];
  minTradeSize?: InputError;
  triggerPrice?: InputError;
  limitPrice?: InputError;
  trades?: InputError;
  fillDelay?: InputError;
  duration?: InputError;
  balance?: InputError;
}

export interface CalculatedOrderForm {
  module: Module;
  isInverted: boolean;
  inputAmount: CalculatedOrderInputAmount;
  outputAmount: CalculatedAmount;
  marketPrice: CalculatedMarketPriceValues;
  trades: CalculatedOrderTrades;
  schedule: CalculatedOrderFormSchedule;
  triggerPrice: CalculatedTriggerPriceValues;
  limitPrice: CalculatedLimitPriceValues;
  minOutputAmountTotal: CalculatedAmount;
  tradePrice: CalculatedAmount;
  fees: CalculatedOrderFees;
  values: CalculatedOrderValues;
  errors: CalculatedOrderFormErrors;
  isReady: boolean;
  canSubmit: boolean;
}